extracted context out to modealProvider
