+++ src/app/calendar.tsx
@@ -1,34 +1,14 @@
 
 "use client"
-import React, { useState } from "react"
+import React from "react"
 import Week from "./week"
 import Modal from "../components/modal/modal"
 import { Deadline } from "@prisma/client"
 import { addDays, differenceInCalendarWeeks } from "date-fns"
-
-export const Context = React.createContext({ openModal: (d: { date: Date, deadlines: (Deadline | null)[] }) => { }, closeModal: () => { }, showModal: false, showCover: false })
+import { ModalProvider } from "@/components/modal/modalProvider"
 
 export default function Calendar({ startDate, semesterStart, weeks, deadlines }: { startDate: Date, semesterStart: Date, weeks: number, deadlines: (Deadline | null)[][] }) {
-  const [showModal, setShowModal] = useState(false)
-  const [showCover, setshowCover] = useState(false)
-
-  const [modalDeadlines, setModalDeadline] = useState<{ date: Date, deadlines: (Deadline | null)[] }>({ date: new Date(), deadlines: [] })
-
-
-  function openModal(d: { date: Date, deadlines: (Deadline | null)[] }) {
-    setModalDeadline(d)
-    setshowCover(true)
-    setTimeout(() => setShowModal(true), 10);
-    document.body.classList.add("modalOpen")
-  }
-
-  function closeModal() {
-    setShowModal(false);
-    setTimeout(() => setshowCover(false), 300);
-    document.body.classList.remove("modalOpen");
-  }
-
-
+  
   const rows = []
   for (var i = 0; i < weeks; i++) {
     const dateOfWeek = addDays(startDate, 7 * i)
@@ -47,21 +27,19 @@ export default function Calendar({ startDate, semesterStart, weeks, deadlines }:
   }
 
   return <>
-    <Context.Provider value={{ openModal, closeModal, showCover, showModal }}>
-      <div className="calendar">
-        <div className="week flex sticky top-0 py-2 glass">
-          {["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"].map((day, i) => (
-            <div className="w-full text-center" key={i}>{day}</div>
+    <div className="calendar">
+      <div className="week flex sticky top-0 py-2 glass">
+        {["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"].map((day, i) => (
+          <div className="w-full text-center" key={i}>{day}</div>
 
-          ))}
-        </div>
+        ))}
+      </div>
+      <ModalProvider>
         {rows}
 
-        <Modal semStart={semesterStart} deadlines={modalDeadlines.deadlines} today={modalDeadlines.date} />
-      </div>
-
-    </Context.Provider>
-
+        <Modal semStart={semesterStart}/>
+      </ModalProvider>
+    </div>
   </>
 
 }
